+++ cmd/dosa/client_test.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2017 Uber Technologies, Inc.
+// Copyright (c) 2018 Uber Technologies, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -22,7 +22,6 @@ package main
 
 import (
 	"context"
-	"reflect"
 	"testing"
 
 	"github.com/golang/mock/gomock"
@@ -78,153 +77,4 @@ func TestClient_Initialize(t *testing.T) {
 	// success case
 	c2 := dosa.NewClient(reg, nullConnector)
 	assert.NoError(t, c2.Initialize(ctx))
-}
-
-func TestClient_Read(t *testing.T) {
-	reg, _ := newSimpleRegistrar(scope, namePrefix, table)
-	fieldsToRead := []string{"ID", "Email"}
-	results := map[string]dosa.FieldValue{
-		"id":    int64(2),
-		"name":  "bar",
-		"email": "bar@email.com",
-	}
-
-	// success case
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	mockConn := mocks.NewMockConnector(ctrl)
-	mockConn.EXPECT().CheckSchema(ctx, gomock.Any(), gomock.Any(), gomock.Any()).Return(int32(1), nil).AnyTimes()
-	mockConn.EXPECT().Read(ctx, gomock.Any(), gomock.Any(), gomock.Any()).
-		Do(func(_ context.Context, _ *dosa.EntityInfo, columnValues map[string]dosa.FieldValue, columnsToRead []string) {
-			assert.Equal(t, dosa.FieldValue(int64(10)), columnValues["id"])
-			assert.Equal(t, []string{"id", "email"}, columnsToRead)
-		}).Return(results, nil).MinTimes(1)
-	c := newShellQueryClient(reg, mockConn)
-	assert.NoError(t, c.Initialize(ctx))
-	fvs, err := c.Read(ctx, []*queryObj{query1}, fieldsToRead, 1)
-	assert.NoError(t, err)
-	assert.Equal(t, 1, len(fvs))
-	assert.Equal(t, results["id"], fvs[0]["ID"])
-	assert.Equal(t, results["name"], fvs[0]["Name"])
-	assert.Equal(t, results["email"], fvs[0]["Email"])
-
-	// error in query, input non-supported operators
-	fvs, err = c.Read(ctx, []*queryObj{query2}, fieldsToRead, 1)
-	assert.Nil(t, fvs)
-	assert.Contains(t, err.Error(), "wrong operator used for read")
-
-	// error in column name converting
-	fvs, err = c.Read(ctx, []*queryObj{query1}, []string{"badcol"}, 1)
-	assert.Nil(t, fvs)
-	assert.Contains(t, err.Error(), "badcol")
-}
-
-func TestClient_Range(t *testing.T) {
-	reg, _ := newSimpleRegistrar(scope, namePrefix, table)
-	fieldsToRead := []string{"ID", "Email"}
-	results := map[string]dosa.FieldValue{
-		"id":    int64(2),
-		"name":  "bar",
-		"email": "bar@email.com",
-	}
-
-	// success case
-	resLimit := 10
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	mockConn := mocks.NewMockConnector(ctrl)
-	mockConn.EXPECT().CheckSchema(ctx, gomock.Any(), gomock.Any(), gomock.Any()).Return(int32(1), nil).AnyTimes()
-	mockConn.EXPECT().Range(ctx, gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
-		Do(func(_ context.Context, _ *dosa.EntityInfo, _ map[string][]*dosa.Condition, minimumFields []string, token string, limit int) {
-			assert.Equal(t, "", token)
-			assert.Equal(t, []string{"id", "email"}, minimumFields)
-			assert.Equal(t, resLimit, limit)
-		}).Return([]map[string]dosa.FieldValue{results}, "", nil).MinTimes(1)
-	c := newShellQueryClient(reg, mockConn)
-	assert.NoError(t, c.Initialize(ctx))
-	fvs, err := c.Range(ctx, []*queryObj{query1, query2}, fieldsToRead, resLimit)
-	assert.NoError(t, err)
-	assert.NotNil(t, fvs)
-	assert.Equal(t, 1, len(fvs))
-	assert.Equal(t, results["id"], fvs[0]["ID"])
-	assert.Equal(t, results["name"], fvs[0]["Name"])
-	assert.Equal(t, results["email"], fvs[0]["Email"])
-
-	// error in query, input non-supported operators
-	fvs, err = c.Range(ctx, []*queryObj{query3}, fieldsToRead, resLimit)
-	assert.Nil(t, fvs)
-	assert.Contains(t, err.Error(), "wrong operator used for range")
-
-	// error in column name converting
-	fvs, err = c.Range(ctx, []*queryObj{query1}, []string{"badcol"}, resLimit)
-	assert.Nil(t, fvs)
-	assert.Contains(t, err.Error(), "badcol")
-}
-
-func TestClient_ColToFieldName(t *testing.T) {
-	colToField := map[string]string{
-		"id":    "ID",
-		"name":  "Name",
-		"email": "Email",
-	}
-
-	rowsCol := []map[string]dosa.FieldValue{
-		{
-			"id":   dosa.FieldValue(int(10)),
-			"name": dosa.FieldValue("foo"),
-		},
-		// contains column "address" which not defined in struct
-		{
-			"id":      dosa.FieldValue(int(20)),
-			"address": dosa.FieldValue("mars"),
-		},
-	}
-
-	expRowsField := []map[string]dosa.FieldValue{
-		{
-			"ID":   dosa.FieldValue(int(10)),
-			"Name": dosa.FieldValue("foo"),
-		},
-		// value of column "address" should not be returned
-		{
-			"ID": dosa.FieldValue(int(20)),
-		},
-	}
-
-	rowsField := convertColToField(rowsCol, colToField)
-	assert.True(t, reflect.DeepEqual(expRowsField, rowsField))
-
-}
-
-func TestClient_BuildReadArgs(t *testing.T) {
-	// success case
-	args, err := buildReadArgs([]*queryObj{query1})
-	assert.NotNil(t, args)
-	assert.NoError(t, err)
-	fv, ok := args["id"]
-	assert.True(t, ok)
-	assert.Equal(t, dosa.FieldValue(int64(10)), fv)
-
-	// fail case, input non-supported operator
-	args, err = buildReadArgs([]*queryObj{query2})
-	assert.Nil(t, args)
-	assert.Contains(t, err.Error(), "wrong operator used for read")
-}
-
-func TestClient_BuildRangeOp(t *testing.T) {
-	limit := 1
-
-	// success case
-	rop, err := buildRangeOp([]*queryObj{query1, query2}, limit)
-	assert.NotNil(t, rop)
-	assert.NoError(t, err)
-	assert.Equal(t, limit, rop.LimitRows())
-	conditions := rop.Conditions()
-	assert.Len(t, conditions, 1)
-	assert.Len(t, conditions["ID"], 2)
-
-	// fail case, input non-supported operator
-	rop, err = buildRangeOp([]*queryObj{query3}, limit)
-	assert.Nil(t, rop)
-	assert.Contains(t, err.Error(), "wrong operator used for range")
 }
